backend/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose
python-multipart

